+++ terra/src/main/scala/com/dafttech/terra/game/world/entities/models/EntityThrown.java
@@ -20,7 +20,7 @@ public abstract class EntityThrown extends Entity {
     }
 
     @Override
-    public float getInAirFriction() {
-        return 0.025f;
+    public double getInAirFriction() {
+        return 0.025;
     }
 }
